Deduplicate variant matching
